requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
aiosqlite==0.19.0
apscheduler==3.10.4
fastapi>=0.109.0
haversine==2.8.0
itsdangerous>=2.1.0
jinja2>=3.1.0
matplotlib
openpyxl
pandas
pdfplumber
python-dotenv==1.0.0
python-multipart>=0.0.6
sqlalchemy==2.0.23
uvicorn[standard]>=0.27.0
